dev-1.3.2.3
Snails no longer move or attack when the session is paused.
HOPEFULLY SNAILS CAN NO LONGER GO INVISIBLE. IT BETTER WORK THIS TIME, I FIXED THIS LIKE A BILLION TIMES ALREADY. I CANT HANDLE THIS ANYMOREEE.

Added Secret Life config: players_drop_task_on_death=false, which controls if the task book is dropped when keep inventory is off.
You can no longer use animal disguise on snails or trivia bots.
Astral projection now puts you into the gamemode you started in, not in survival.
Using the Super Punch power, you can no longer ride an entity that already had a passenger on it, and you now CAN ride on entities that have a vehicle, but no passengers.

**Added Simple Life** (summons the wandering traders with the custom trades, up to a max of three on the world)

+++ src/main/java/net/mat0u5/lifeseries/config/DatapackManager.java
@@ -15,6 +15,7 @@ public class DatapackManager {
         try {
             for (SeriesList series : SeriesList.getAllImplemented()) {
                 String datapackName = SeriesList.getDatapackName(series);
+                if (datapackName == null) continue;
                 Path datapackPath = datapackFolder.resolve(datapackName);
                 if (Files.exists(datapackPath) && Files.isRegularFile(datapackPath)) {
                     Files.delete(datapackPath);
